refactor(llm): extract provider config + reasoning helpers to _provider_config
Moves the provider routing config (PROVIDER_MODEL_PREFIX, PROVIDER_ENV_KEY),
the reasoning-model output-token floor/cap constants, and the three
reasoning-detection helpers (_openai_model_id,
_is_openai_reasoning_style_model, _supports_thinking) out of
provider.py into amx/llm/_provider_config.py.

provider.py re-exports each public name so historical imports continue
to work. provider.py: 1504 -> 1401 LOC (-103, -7%).
